Api/city

+++ internal/api/handler/response/city.go
@@ -1,6 +1,6 @@
 package response
 
 type ICityResponse struct {
-	Id   int    `json:"_id"`
+	Id   int64  `json:"_id"`
 	Name string `json:"name"`
 }
